refactor: cluster thunks
