+++ src/redux/thunks/cluster/stopClusterConnection.ts
@@ -1,0 +1,17 @@
+import {createAsyncThunk} from '@reduxjs/toolkit';
+
+import {setClusterProxyPort} from '@redux/reducers/appConfig';
+import {disconnectFromCluster} from '@redux/services/clusterResourceWatcher';
+
+import {AppDispatch} from '@shared/models/appDispatch';
+import {RootState} from '@shared/models/rootState';
+import {closeKubectlProxy} from '@shared/utils/commands/kubectl';
+
+export const stopClusterConnection = createAsyncThunk<void, undefined, {dispatch: AppDispatch; state: RootState}>(
+  'main/stopClusterConnection',
+  async (payload, thunkAPI) => {
+    thunkAPI.dispatch(setClusterProxyPort(undefined));
+    disconnectFromCluster();
+    closeKubectlProxy();
+  }
+);
